+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Deprecated
 
 ### Removed
-
+- Dependency to deprecated oslc4j-json4j-provider
 ### Fixed
 
 - Client now picks the correct ResponseInfo object when an OSLC Query response contains multiple ResponseInfo objects.